#184 Turn dynamic strings on by default for LMDB too.
This also takes the opportunity to run many of the `test_arctic.py` tests against the LMDB backend too, whose test coverage was pretty non-existent. The LMDB backed tests execute extremely quickly so build time is not a concern.
